requirements.txt
pypdf
pyxdg
